# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Street Burger API",
        "description": "Street Burger Mobile Application Backend API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservations/add": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a new reservation",
                "operationId": "App\\Http\\Controllers\\ReservationController::addReservation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "guest_count",
                                    "reservation_date",
                                    "reservation_time"
                                ],
                                "properties": {
                                    "guest_count": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-20"
                                    },
                                    "reservation_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "19:00:00"
                                    },
                                    "special_requests": {
                                        "type": "string",
                                        "example": "Window seat please"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/getByPhone/{phoneNumber}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get reservations by phone number",
                "operationId": "App\\Http\\Controllers\\ReservationController::getByPhone",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reservations"
                    }
                }
            }
        },
        "/api/reservations/getById/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get reservation by ID",
                "operationId": "App\\Http\\Controllers\\ReservationController::getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation details"
                    }
                }
            }
        },
        "/api/reviews/add/{phoneNumber}": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Add a new review",
                "operationId": "App\\Http\\Controllers\\ReviewController::addReview",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "comment"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Great burger!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review added"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/rwdpts/addrwdpts": {
            "post": {
                "tags": [
                    "Rewards"
                ],
                "summary": "Add reward points",
                "operationId": "App\\Http\\Controllers\\RewardController::addPoints",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "points",
                                    "description"
                                ],
                                "properties": {
                                    "points": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Order bonus"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Points added"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rwdpts/getrwdpts": {
            "get": {
                "tags": [
                    "Rewards"
                ],
                "summary": "Get user reward points",
                "operationId": "App\\Http\\Controllers\\RewardController::getPoints",
                "responses": {
                    "200": {
                        "description": "Total points and transaction history"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}